Typing and doc improvements in the munge.
